requirements.txt
aiogram==3.*
asyncpg==0.29.*
python-dotenv==1.*
ijson==3.*